Squashed commit of the following:
commit 99acb0af46d38c9e2dab64fa185a098c660660f9
Author: LordHajAli <anajar13750@gmail.com>
Date:   Tue May 2 23:38:40 2023 +0330

    endGame

commit 64bbdb56ad945f38dd63f1cf49541d374f871d7a
Author: LordHajAli <anajar13750@gmail.com>
Date:   Tue May 2 23:01:13 2023 +0330

    end Game , attacking each other

commit a5891022a48f45c4a4a9e3f93c9706bdd9fafeb3
Merge: 89d9c39 5b7a9a7
Author: LordHajAli <anajar13750@gmail.com>
Date:   Sun Apr 30 23:13:22 2023 +0330

    Merge branch 'main' of https://github.com/advanced-programming-sut-2023/project-group-07

commit 89d9c3976929d78d64f42f3509102729e3cd81a6
Author: LordHajAli <anajar13750@gmail.com>
Date:   Sun Apr 30 22:33:25 2023 +0330

    bugs fixed

commit 7fd4429b834e5ebc17a8b1c43548446ccda2a3aa
Author: LordHajAli <anajar13750@gmail.com>
Date:   Sun Apr 30 20:37:23 2023 +0330

    bugs fixed

+++ Stronghold crusader/controller/TradeMenuController.java
@@ -11,43 +11,46 @@ import java.util.Locale;
 public class TradeMenuController {
 
     public ArrayList<TradeRequest> getNotSeenRequests() {
-        Government government =  Controller.currentGame.getCurrentGovernment();
+        Government government = Controller.currentGame.getCurrentGovernment();
         return TradeRequest.getNotSeenRequests(government);
     }
 
     public void setRequestsShown(ArrayList<TradeRequest> requests) {
-        for(TradeRequest tradeRequest : requests){
+        for (TradeRequest tradeRequest : requests) {
             tradeRequest.setHaveBeenShownTrue();
         }
     }
 
     public ArrayList<TradeRequest> getAvailableTrades() {
-        Government government =  Controller.currentGame.getCurrentGovernment();
+        Government government = Controller.currentGame.getCurrentGovernment();
         return TradeRequest.getAvailableRequests(government);
     }
 
     public ArrayList<TradeRequest> getTradeHistory() {
-        Government government =  Controller.currentGame.getCurrentGovernment();
+        Government government = Controller.currentGame.getCurrentGovernment();
         return TradeRequest.getRelatedTradeHistory(government);
     }
 
     public Messages acceptTrade(int id, String message) {
-        Government receiver =  Controller.currentGame.getCurrentGovernment();
+        Government receiver = Controller.currentGame.getCurrentGovernment();
         ArrayList<TradeRequest> availableTrades = TradeRequest.getAvailableRequests(receiver);
         TradeRequest targetRequest = null;
-        for(TradeRequest request : availableTrades){
-            if (request.getId() == id){
+        for (TradeRequest request : availableTrades) {
+            if (request.getId() == id) {
                 targetRequest = request;
                 break;
             }
         }
-        if (targetRequest==null) return Messages.INVALID_ID;
+        if (targetRequest == null)
+            return Messages.INVALID_ID;
         Government requester = targetRequest.requester();
-        double price = targetRequest.price();
-        if (requester.getGold() < price) return Messages.POOR_REQUESTER;
+        int price = targetRequest.price();
+        if (requester.getGold() < price)
+            return Messages.POOR_REQUESTER;
         int amount = targetRequest.amount();
         Resources resource = targetRequest.resource();
-        if (receiver.getResourceAmount(resource) < amount) return Messages.POOR_RECEIVER;
+        if (receiver.getResourceAmount(resource) < amount)
+            return Messages.POOR_RECEIVER;
         receiver.changeResources(resource, -amount);
         receiver.changeGold(price);
         requester.changeResources(resource, amount);
@@ -57,31 +60,37 @@ public class TradeMenuController {
     }
 
     public Messages rejectTrade(int id) {
-        Government receiver =  Controller.currentGame.getCurrentGovernment();
+        Government receiver = Controller.currentGame.getCurrentGovernment();
         ArrayList<TradeRequest> availableTrades = TradeRequest.getAvailableRequests(receiver);
         TradeRequest targetRequest = null;
-        for(TradeRequest request : availableTrades){
-            if (request.getId() == id){
+        for (TradeRequest request : availableTrades) {
+            if (request.getId() == id) {
                 targetRequest = request;
                 break;
             }
         }
-        if (targetRequest==null) return Messages.INVALID_ID;
+        if (targetRequest == null)
+            return Messages.INVALID_ID;
         targetRequest.reject();
         return Messages.REJECT_TRADE_SUCCESSFUL;
     }
 
-    public Messages requestTrade(String typeName, String colorString, String message, int amount, double price) {
-        if (price < 0) return Messages.NEGATIVE_PRICE;
+    public Messages requestTrade(String typeName, String colorString, String message, int amount, int price) {
+        if (price < 0)
+            return Messages.NEGATIVE_PRICE;
         Resources resource = Resources.getResource(typeName);
-        if (resource == null) return Messages.INVALID_RESOURCE;
+        if (resource == null)
+            return Messages.INVALID_RESOURCE;
         LordColor color = LordColor.getColorByName(colorString);
-        if (color == null) return Messages.INVALID_COLOR;
+        if (color == null)
+            return Messages.INVALID_COLOR;
         Government receiver = Controller.currentGame.getGovernmentByColor(color);
-        if (receiver == null) return Messages.NO_LORD_WITH_THIS_COLOR;
+        if (receiver == null)
+            return Messages.NO_LORD_WITH_THIS_COLOR;
         Government requester = Controller.currentGame.getCurrentGovernment();
-        if (receiver.equals(requester)) return Messages.REQUEST_YOURSELF;
-        TradeRequest.makeATradeRequest(price,resource,amount,requester,receiver,message);
+        if (receiver.equals(requester))
+            return Messages.REQUEST_YOURSELF;
+        TradeRequest.makeATradeRequest(price, resource, amount, requester, receiver, message);
         return Messages.REQUEST_TRADE_SUCCESSFUL;
 
     }
